fix: improve API safety
* Remove unsound TryFrom<*mut YR_RULES> for Rules

This trait implementation was unsound: A user could trivially pass a
null or dangling pointer to it, and call any of the safe methods on
Rules to trigger a segfault.

* Fix lifetimes of the scan_x functions.

The previous lifetimes were totally broken: They allowed the caller of
the function to specify the lifetimes they wanted. It was possible to
get a Rule<'static> out of a Rules<'a>!

Prevent this by binding the output Rules lifetime to self.