feat: rework configuration to remove enum
Each configuration that is currently handled works with a u32. However,
another configuration that needs to be added works with a u64. This
makes the current API unwieldy.

A simple solution to have a stronger API is to explicit each element
with its own get/set functions. This way, the right type for this config
can be used.

This is feasible because only 3 parameters are configurable for the
moment (and a fourth will be added).